+++ fluent-intl-polyfill/bundle_config.js
@@ -1,0 +1,9 @@
+import nodeResolve from 'rollup-plugin-node-resolve';
+
+export default {
+  context: 'this',
+  plugins: [
+    nodeResolve(),
+  ]
+};
+